oops all int16_ts

+++ radio.ino
@@ -31,10 +31,10 @@ void OnTxDone( void );
 void OnTxTimeout( void );
 void OnRxDone( uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr );
 
-int16_t demod(int16_t rssi);
-int16_t scaleSignal(int16_t dist, int8_t range);
-int16_t timeReturn(void);
-int16_t scaleTime(int16_t time, int16_t range);
+int demod(int16_t rssi);
+int scaleSignal(int dist, int range);
+int timeReturn(void);
+int scaleTime(int time, int range);
 
 typedef enum
 {
@@ -127,7 +127,7 @@ void OnRxDone( uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr )
     Radio.Sleep( );
 
     // Hoping this will work too. Should just give packet time data for us to test and create distance vs time approximations
-    int16_t receiveTime = timeReturn();
+    int receiveTime = timeReturn();
     Serial.printf("\r\nreceived packet \"%s\" with Rssi %d , length %d , packet time %d\r\n",rxpacket,Rssi,rxSize, receiveTime);
     Serial.println("wait to send next packet");
 
@@ -137,39 +137,39 @@ void OnRxDone( uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr )
 /*I'm using the absolute value of the dbm because at -100 dbm it will
 be converted to approximately 10MW and -5 dbm will be 3mW. IMO nice because
 a farther distance is just a bigger number.*/
-int16_t demod(int16_t rssi)
+int demod(int16_t rssi)
 {
-  uint16_t dist = ceil((pow(10, (abs(rssi)/10))));
+  int dist = ceil((pow(10, (abs(rssi)/10))));
  // Converts signal strenght from dBm but below a certain margin will just put it to 0 ie right beside each other
 
   return dist;
 }
 
 //Since our maximum value was around -100 dbm, I will put the scale range to a bit above 10MW
-int16_t scaleSignal(int16_t dist, int8_t range)
+int scaleSignal(int dist, int range)
 {
   //dist will be the value obtained from demod
   //range will be the input range for the different UI display types
   //ie for drawCircles its 0-117 so I will put 117 as range
-  int16_t max = 100000000;
+  int max = 100000000;
 
   float ratio = dist/max;
 
-  int16_t scaled = floor(ratio*range);
+  int scaled = floor(ratio*range);
 
   return scaled;
 }
 
 //Returns the time it's been since last this function ran
-int16_t timeReturn(void)
+int timeReturn(void)
 {
   uint8_t tgl = 0;
-  uint16_t newTime = micros();
-  uint16_t oldTime;
+  int newTime = micros();
+  int oldTime;
 
   if(tgl != 0)
   {
-    int16_t interval = newTime - oldTime;
+    int interval = newTime - oldTime;
 
     return interval;
   }
@@ -182,13 +182,13 @@ int16_t timeReturn(void)
   oldTime = newTime;
 }
 
-int16_t scaleTime(int16_t time, int16_t range)
+int scaleTime(int time, int range)
 {
-  int16_t maxTime = 100; //placeholder that assumes farthest distance is 100us travel time
-  int16_t ratio = (time-1000000 /*1 second delay*/)/maxTime;
+  int maxTime = 100; //placeholder that assumes farthest distance is 100us travel time
+  int ratio = (time-1000000 /*1 second delay*/)/maxTime;
 
   // range will be maximum input for draw functions
-  int16_t dist = (ratio*range);
+  int dist = (ratio*range);
 }
